+++ modules/scanners.nix
@@ -1,0 +1,10 @@
+{
+  nixosModule = { user, pkgs, ... }: {
+    hardware.sane.enable = true;
+
+    users.users.${user}.extraGroups = [ "scanner" ];
+
+    environment.systemPackages =
+      builtins.attrValues { inherit (pkgs.gnome) simple-scan; };
+  };
+}